Clarify the role of open source license and CSL
I'm not sure what was intended, but would it help to have a clarification that in the case of confusion or conflict between the MIT or other listed OSI license terms and the CSL, that the CSL applies? I'm thinking of a situation where someone includes code in the specification repo with a random OSI approved license identified (e.g. in the header) that causes confusion about what applies.