Update ..Getting Started.md
In the CLA section, the term "Participant" was capitalized but undefined. It's also not defined in the spec itself (or I would reference that definition) and so I have reworded that section to read:  "1. **CLA bot.** Enable a CLA Bot, such as EasyCLA or cla-bot, to require a **Community Specification Contributor License Agreement** signed (either by an individual contributor or by a contributor's employer, which CLA covers the employed contributor) and in place prior to making any contribution."